REFACTOR: 모든 동적 쿼리 디폴트 정렬 createdDate desc 로 수정(쿼리스트링에서 받지 않음)

+++ src/main/java/couch/camping/domain/review/service/ReviewServiceImpl.java
@@ -102,6 +102,11 @@ public class ReviewServiceImpl implements ReviewService {
                     .orElseThrow(() -> {
                         throw new CustomException(ErrorCode.NOT_FOUND_REVIEW, "리뷰 ID 에 맞는 리뷰가 없습니다.");
                     });
+
+            if (findReview.getMember() != member) {
+                throw new CustomException(ErrorCode.FORBIDDEN_MEMBER, "해당 회원의 리뷰가 아닙니다.");
+            }
+
             findReview.getCamp().decreaseRate(findReview.getRate());
             reviewRepository.deleteById(reviewId);
         } catch (EmptyResultDataAccessException e) {
